Made a cleaner test file to show how to use the extraction library
