+++ ransac_primitives/CMakeLists.txt
@@ -7,15 +7,12 @@ set(CMAKE_CXX_FLAGS "-O4 -Wall -fPIC -std=c++0x")
 
 find_package(catkin REQUIRED)
 # PCL
-#set(PCL_DIR "/usr/share/pcl-1.7")
 find_package(PCL 1.7 REQUIRED)
 include_directories(${PCL_INCLUDE_DIRS})
 link_directories(${PCL_LIBRARY_DIRS})
 add_definitions(${PCL_DEFINITIONS})
 
 # OpenCV
-#include_directories(/opt/ros/hydro/include)
-#link_directories(/opt/ros/hydro/lib)
 find_package(OpenCV REQUIRED)
 
 catkin_package(
@@ -41,8 +38,7 @@ catkin_package(
 add_library(primitive_octree src/primitive_leaf.cpp src/primitive_octree.cpp src/base_primitive.cpp)
 add_library(primitive_extractor src/primitive_visualizer.cpp src/primitive_extractor.cpp)
 add_library(basic_primitives src/plane_primitive.cpp src/sphere_primitive.cpp src/cylinder_primitive.cpp)
-#add_executable(test_primitives src/test_primitives.cpp)
-add_executable(test_find_point src/test_find_point.cpp)
+add_executable(test_extraction src/test_extraction.cpp)
 
 #target_link_libraries(base_primitive primitive_octree ${OpenCV_LIBS} ${PCL_LIBRARIES}) # opencv_core opencv_imgproc opencv_highgui
 #target_link_libraries(plane_primitive primitive_octree ${PCL_LIBRARIES})
@@ -59,5 +55,4 @@ add_executable(test_find_point src/test_find_point.cpp)
 target_link_libraries(basic_primitives primitive_octree ${PCL_LIBRARIES})
 target_link_libraries(primitive_octree ${OpenCV_LIBS} ${PCL_LIBRARIES})
 target_link_libraries(primitive_extractor primitive_octree ${PCL_LIBRARIES})
-#target_link_libraries(test_primitives primitive_extractor basic_primitives ${PCL_LIBRARIES})
-target_link_libraries(test_find_point primitive_extractor basic_primitives ${PCL_LIBRARIES})
+target_link_libraries(test_extraction primitive_extractor basic_primitives ${PCL_LIBRARIES})
